requirements.txt
streamlit~=1.47.0
pandas~=2.3.1
numpy~=2.3.2
matplotlib~=3.10.6
seaborn~=0.13.2
openpyxl~=3.1.5
xlrd~=2.0.2
pytest~=8.4.1
tqdm~=4.67.1
google-auth
google-auth-oauthlib~=1.2.2
google-auth-httplib2~=0.2.0
google-api-python-client~=2.181.0
mysql-connector-python~=9.4.0
plotly~=6.2.0

GitPython~=3.1.45
gitdb~=4.0.12
ipython~=9.4.0
pillow~=11.3.0
pip~=25.2
typing_extensions~=4.14.1
attrs~=25.3.0
protobuf~=6.31.1
pyOpenSSL~=25.1.0
cryptography~=45.0.7
rsa~=4.9.1
pyasn1~=0.6.1
pytz~=2025.2
toml~=0.10.2
requests~=2.32.5
colorama~=0.4.6
smmap~=5.0.2
charset-normalizer~=3.4.2
cffi~=2.0.0
proto-plus~=1.26.1
altair~=5.5.0
jsonschema~=4.25.0
narwhals~=1.48.1
pyarrow~=21.0.0
Jinja2~=3.1.6
packaging~=25.0
referencing~=0.36.2
traitlets~=5.14.3
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.8.3
urllib3~=2.5.0
MarkupSafe~=3.0.2
SQLAlchemy~=2.0.41
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
Pygments~=2.19.2
wcwidth~=0.2.13
pluggy~=1.6.0
iniconfig~=2.1.0
tzdata~=2025.2
cycler~=0.12.1
tornado~=6.5.1
six~=1.17.0
pyparsing~=3.2.3
httplib2~=0.30.0
oauthlib~=3.3.1
blinker~=1.9.0
et_xmlfile~=2.0.0
idna~=3.10
tenacity~=9.1.2
watchdog~=6.0.0
contourpy~=1.3.3
fonttools~=4.59.2
click~=8.2.1
jsonschema-specifications~=2025.4.1
ipykernel~=6.30.0
kiwisolver~=1.4.9
uritemplate~=4.2.0
requests-oauthlib~=2.0.0
typing_extensions~=4.14.1
pyasn1_modules~=0.4.2
et_xmlfile~=2.0.0
reportlab
